Add mysql version layer
