+++ lib/gratan.rb
@@ -11,7 +11,9 @@ module Gratan; end
 require 'gratan/logger'
 require 'gratan/template_helper'
 require 'gratan/client'
-require 'gratan/driver'
+require 'gratan/mysql5'
+require 'gratan/mysql5/client'
+require 'gratan/mysql5/driver'
 require 'gratan/dsl'
 require 'gratan/dsl/validator'
 require 'gratan/dsl/context'
