+++ lib/gratan/mysql5.rb
@@ -1,0 +1,2 @@
+class Gratan::Mysql5
+end
