+++ spec/mysql5/create/create_user_regexp_spec.rb
