+++ spec/mysql5/create/create_user_with_template_spec.rb
