+++ spec/mysql5/misc/require_spec.rb
